First working version
XZCompress will break up a file into a user specified chunk size, compress the chunks separately and write out JSON describing each chunk.
It'll also try each of DEFLATE's compression strategies and use the one that produces the smallest size.
The output JSON file appends data for different files so that it becomes a collated collection of compressed chunk descriptions for say, a whole collection of data rather than one file.
If a file is already in the JSON data, it's record is overwritten.

Tested across a few file types in debug and release.